Difference in -> difference between
When discussing discreet things "distance between" is preferred. "Distance in" is preferred for more abstract or less discrete ideas. For example:
distance between the means
distance between the groups
distance in the height of the groups
distance in the age of the groups
distance between his and her ages.
distance in their ages -> doesn't sound bad, but not as good as "between."
This becomes a bit more problematic with the phrasing "distance in means" where means can mean a discrete thing the, ahem, "average" or it can mean an abstract thing like "the way of doing something." Sure you can get it from context, but the reader who is aware of the difference has to double check every time.